// src/gantt/types/gantt-task-actions.ts
import type { BarTask } from './bar-task'

export type BarMoveAction = 'progress' | 'end' | 'start' | 'move'
export type GanttContentMoveAction =
  | 'mouseenter'
  | 'mouseleave'
  | 'delete'
  | 'dblclick'
  | 'click'
  | 'select'
  | ''
  | BarMoveAction

export type GanttEvent = {
  changedTask?: BarTask
  originalSelectedTask?: BarTask
  action: GanttContentMoveAction
}